refactor: extract profile dialogs from navbar
